refactor: apply PR requested changes
