+++ src/utils/hooks.ts
@@ -4,8 +4,7 @@ import { ROUTINES, useLocalContext } from '@graasp/apps-query-client';
 
 import Uppy, { UploadResult } from '@uppy/core';
 
-import { Settings, SettingsKeys } from '@/@types';
-import { hooks, mutations, notifier } from '@/config/queryClient';
+import { mutations, notifier } from '@/config/queryClient';
 
 import configureUppy from './uppy';
 
@@ -101,38 +100,4 @@ export const useUploadImage = ({
   }, []);
 
   return uppy;
-};
-
-// mapping between Setting names and their data type
-// eslint-disable-next-line @typescript-eslint/ban-types
-type AllSettingsType = {
-  [SettingsKeys.Settings]: Settings;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  [SettingsKeys.File]: any;
-};
-type AllSettingsNameType = SettingsKeys;
-type AllSettingsDataType = AllSettingsType[keyof AllSettingsType];
-export const saveSettings = (
-  name: AllSettingsNameType,
-  newValue: AllSettingsDataType,
-): void => {
-  const { mutate: postAppSetting } = mutations.usePostAppSetting();
-  const { mutate: patchAppSetting } = mutations.usePatchAppSetting();
-  const { data: appSettingsList } = hooks.useAppSettings();
-
-  if (appSettingsList) {
-    const previousSetting = appSettingsList.find((s) => s.name === name);
-    // setting does not exist
-    if (!previousSetting) {
-      postAppSetting({
-        data: newValue,
-        name,
-      });
-    } else {
-      patchAppSetting({
-        id: previousSetting.id,
-        data: newValue,
-      });
-    }
-  }
 };
